+++ mypy/join.py
@@ -756,10 +756,24 @@ def join_similar_callables(t: CallableType, s: CallableType) -> CallableType:
     # TODO in combine_similar_callables also applies here (names and kinds; user metaclasses)
     # The fallback type can be either 'function', 'type', or some user-provided metaclass.
     # The result should always use 'function' as a fallback if either operands are using it.
-    if t.fallback.type.fullname == "builtins.function":
+    #  Based: typing._Callable is the fallback for Callable, but there is also types.FunctionType
+    # There is an issue where `function` doesn't extend typing._Callable
+    #  Additionally, this only happens when union joins are disabled (unbased mode)
+    if t.fallback_name == s.fallback_name:
         fallback = t.fallback
-    else:
+    elif t.fallback_name == "typing._Callable":
+        fallback = t.fallback
+    elif s.fallback_name == "typing._Callable":
         fallback = s.fallback
+    else:
+        joined = join_types(t.fallback, s.fallback)
+        assert isinstance(joined, Instance)
+        fallback = joined
+        if fallback.type.fullname == "builtins.object":
+            from mypy.typeanal import CALLABLE_TYPE
+
+            assert CALLABLE_TYPE
+            fallback = CALLABLE_TYPE
     return t.copy_modified(
         arg_types=arg_types,
         arg_names=combine_arg_names(t, s),
@@ -809,7 +823,7 @@ def combine_similar_callables(t: CallableType, s: CallableType) -> CallableType:
     # TODO what should happen if one fallback is 'type' and the other is a user-provided metaclass?
     # The fallback type can be either 'function', 'type', or some user-provided metaclass.
     # The result should always use 'function' as a fallback if either operands are using it.
-    if t.fallback.type.fullname == "builtins.function":
+    if t.fallback_name in {"builtins.function", "typing._Callable"}:
         fallback = t.fallback
     else:
         fallback = s.fallback